Optimize calculations using single-precision floats
This improves performance by ~30% on my laptop without affecting results
significantly.

+++ loader.py
@@ -56,13 +56,13 @@ def load_mnist_dataset(path, shape=(28 * 28,)):
     """
     train_images = load_mnist_images(f"{path}/train-images-idx3-ubyte")
     train_images = train_images.reshape((60000, *shape))
-    train_images = train_images.astype("float") / 255.0
+    train_images = train_images.astype("float32") / 255.0
     train_labels = load_mnist_labels(f"{path}/train-labels-idx1-ubyte")
 
     print("reading test data...")
     test_images = load_mnist_images(f"{path}/t10k-images-idx3-ubyte")
     test_images = test_images.reshape((10000, *shape))
-    test_images = test_images.astype("float") / 255.0
+    test_images = test_images.astype("float32") / 255.0
     test_labels = load_mnist_labels(f"{path}/t10k-labels-idx1-ubyte")
 
     return (train_images, train_labels, test_images, test_labels)
